move config helpers to separate file, use everywhere
