+++ config.py
@@ -1,0 +1,21 @@
+from aqt import mw
+
+def gc(arg="", fail=False):
+    conf = mw.addonManager.getConfig(__name__)
+    if conf:
+        if arg:
+            return conf.get(arg, fail)
+        else:
+            return conf
+    return fail
+
+
+def dc(arg="", fail=""):
+    addon = mw.addonManager.addonFromModule(__name__)
+    conf = mw.addonManager.addonConfigDefaults(addon)
+    if conf:
+        if arg:
+            return conf.get(arg, fail)
+        else:
+            return conf
+    return fail
